Fibonacci 98.2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 	long i;
 	int n = 98;
 	long t1 = 1, t2 = 2;
-	unsigned long long nextTerm = t1 + t2;
+	unsigned long nextTerm = t1 + t2;
 
 	printf("%li, %li, ", t1, t2);
 
@@ -17,13 +17,13 @@ int main(void)
 	{
 		if (i < n)
 		{
-			printf("%llu, ", nextTerm);
+			printf("%lu, ", nextTerm);
 			t1 = t2;
 			t2 = nextTerm;
 			nextTerm = t1 + t2;
 		}
 		else
-			printf("%llu", nextTerm);
+			printf("%lu", nextTerm);
 	}
 	printf("\n");
 	return (0);
